+++ src/models/banner.ts
@@ -1,0 +1,7 @@
+export interface Banner {
+  bannerId: number
+  pic: string
+  targetId: number
+  targetType: number
+  typeTitle: string
+}