+++ src/router/meta.d.ts
@@ -1,0 +1,12 @@
+import 'vue-router'
+
+export {}
+
+declare module 'vue-router' {
+  interface RouteMeta {
+    keepAlive?: boolean
+    menu?: string
+    requiresApiHost?: boolean
+    title: string
+  }
+}